♻️ [refactor/withdraw] CKPS-17 users 모델에서 withdrawal 구조 분리
