+++ apps/users/models/users.py
@@ -1,12 +1,7 @@
-from datetime import timedelta
-from typing import Any
-
 from django.contrib.auth.base_user import AbstractBaseUser
 from django.db import models
-from django.utils import timezone
 
 from apps.core.models import TimeStampedModel
-from apps.users.utils.reason_choices import WithdrawalReason
 
 
 class GenderChoices(models.TextChoices):
@@ -29,23 +24,3 @@ class User(TimeStampedModel, AbstractBaseUser):
 
     class Meta:
         db_table = "users"
-
-
-class Withdrawal(TimeStampedModel):
-    user = models.ForeignKey(
-        User, on_delete=models.SET_NULL, null=True, related_name="withdrawals", db_column="user_id"
-    )
-    reason = models.CharField(
-        max_length=100,
-        choices=WithdrawalReason.choices,
-    )
-    reason_detail = models.CharField(max_length=500, blank=True, default="")
-    due_date = models.DateField()
-
-    class Meta:
-        db_table = "withdrawals"
-
-    def save(self, *args: Any, **kwargs: Any) -> None:
-        if not self.due_date:
-            self.due_date = timezone.now().date() + timedelta(days=14)
-        super().save(*args, **kwargs)
